+++ test/visualTest.ts
@@ -804,117 +804,218 @@ describe("SankeyDiagram", () => {
         });
     });
 
-    function timeout(ms: number) {
-        return new Promise(resolve => setTimeout(resolve, ms));
-    }
-
-    const defaultWaitForRender: number = 500;
-
-    describe("Keyboard Navigation check", () =>{
-        it("links should have attributes tabindex=0, role=option, aria-label is not null, and aria-selected=false", (done) => {
+    describe("Keyboard Navigation tests:", () => {
+        it("links should have attributes tabindex>0, role=option, aria-label is not null, and aria-selected=false", (done: DoneFn) => {
             visualBuilder.updateRenderTimeout(dataView, () => {
-                // defaults
-                const someColor: string = "#000000";
-                const fontSize: number = 12;
-                const unit: number = 0;
-
-                visualBuilder.instance.getFormattingModel();
-
-                let nodes = visualBuilder.linkElements;
-                nodes.forEach((el: Element) => {
+                let links = visualBuilder.linkElements;
+                links.forEach((el: Element) => {
                     expect(el.getAttribute("role")).toBe("option");
-                    expect(el.getAttribute("tabindex")).toBe("0");
+                    expect(el.getAttribute("tabindex")).toBeGreaterThanOrEqual(1);
                     expect(el.getAttribute("aria-selected")).toBe("false");
                     expect(el.getAttribute("aria-label")).not.toBeNull();
                 });
                 done();
-            },);
+            });
+        });
+
+        it("nodes should have attributes tabindex>0, role=option, aria-label is not null, and aria-selected=false", (done: DoneFn) => {
+            visualBuilder.updateRenderTimeout(dataView, () => {
+                let nodeRects = visualBuilder.nodeRectElements;
+                nodeRects.forEach((el: Element) => {
+                    expect(el.getAttribute("role")).toBe("option");
+                    expect(el.getAttribute("tabindex")).toBeGreaterThanOrEqual(1);
+                    expect(el.getAttribute("aria-selected")).toBe("false");
+                    expect(el.getAttribute("aria-label")).not.toBeNull();
+                });
+                done();
+            });
         });
 
         it("enter toggles the correct slice", (done: DoneFn) => {
             const enterEvent = new KeyboardEvent("keydown", { code: "Enter", bubbles: true });
             visualBuilder.updateRenderTimeout(
-                dataView,
-                    async () => {
-                        visualBuilder.linkElements[0].dispatchEvent(enterEvent);
-                        await timeout(defaultWaitForRender);
-                        expect(visualBuilder.linkElements[0].getAttribute("aria-selected")).toBe("true");
-                        for (const slice of visualBuilder.linkElements) {
-                            if (slice !== visualBuilder.linkElements[0]) {
-                                expect(slice.getAttribute("aria-selected")).toBe("false");
-                            }
-                        }
+                dataView, () => {
+                        const links: HTMLElement[] = [...visualBuilder.linkElements];
 
-                        visualBuilder.linkElements[0].dispatchEvent(enterEvent);
-                        await timeout(defaultWaitForRender);
-                        for (const slice of visualBuilder.linkElements) {
-                            expect(slice.getAttribute("aria-selected")).toBe("false");
-                        }
+                        links[0].dispatchEvent(enterEvent);
 
+                        expect(links[0].getAttribute("aria-selected")).toBe("true");
+
+                        const otherLinks: HTMLElement[] = links.slice(1);
+                        otherLinks.forEach((link: HTMLElement) => {
+                            expect(link.getAttribute("aria-selected")).toBe("false");
+                        })
+
+                        links[0].dispatchEvent(enterEvent);
+
+                        links.forEach((element: Element) => {
+                            expect(element.getAttribute("aria-selected")).toBe("false");
+                        });
                         done();
                     },
                 2,
-                );
+            );
+        });
+        
+        it("space toggles the correct slice", (done: DoneFn) => {
+            const spaceEvent = new KeyboardEvent("keydown", { code: "Space", bubbles: true });
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                    () => {
+                        const links: HTMLElement[] = [...visualBuilder.linkElements];
+
+                        links[0].dispatchEvent(spaceEvent);
+
+                        expect(links[0].getAttribute("aria-selected")).toBe("true");
+                        
+                        const otherLinks: HTMLElement[] = links.slice(1);
+                        otherLinks.forEach((link: HTMLElement) => {
+                            expect(link.getAttribute("aria-selected")).toBe("false");
+                        });
+
+                        links[0].dispatchEvent(spaceEvent);
+
+                        links.forEach((element: HTMLElement) => {
+                            expect(element.getAttribute("aria-selected")).toBe("false");
+                        });
+                        done();
+                    },
+                2,
+            );
+        });
+        
+        it("tab between slices works", (done: DoneFn) => {
+            const tabEvent = new KeyboardEvent("keydown", { code: "Tab", bubbles: true });
+            const enterEvent = new KeyboardEvent("keydown", { code: "Enter", bubbles: true });
+            visualBuilder.updateRenderTimeout(
+                dataView,
+                    () => {
+                        const links: HTMLElement[] = [...visualBuilder.linkElements];
+
+                        links[0].dispatchEvent(enterEvent);
+                        expect(links[0].getAttribute("aria-selected")).toBe("true");
+
+                        const otherLinks: HTMLElement[] = links.slice(1);
+                        otherLinks.forEach((link: HTMLElement) => {
+                            expect(link.getAttribute("aria-selected")).toBe("false");
+                        });
+
+                        visualBuilder.element.dispatchEvent(tabEvent);
+
+                        links[1].dispatchEvent(enterEvent);
+                        expect(links[1].getAttribute("aria-selected")).toBe("true");
+
+                        links.splice(1, 1);
+                        links.forEach((link: HTMLElement) => {
+                            expect(link.getAttribute("aria-selected")).toBe("false");
+                        });
+                        done();
+                    },
+                2,
+            );
         });
     });
-    
-    it("space toggles the correct slice", (done: DoneFn) => {
-        const spaceEvent = new KeyboardEvent("keydown", { code: "Space", bubbles: true });
-        visualBuilder.updateRenderTimeout(
-            dataView,
-                async () => {
-                    visualBuilder.linkElements[0].dispatchEvent(spaceEvent);
-                    await timeout(defaultWaitForRender);
-                    expect(visualBuilder.linkElements[0].getAttribute("aria-selected")).toBe("true");
-                    for (const slice of visualBuilder.linkElements) {
-                        if (slice !== visualBuilder.linkElements[0]) {
-                            expect(slice.getAttribute("aria-selected")).toBe("false");
-                        }
-                    }
 
-                    visualBuilder.linkElements[0].dispatchEvent(spaceEvent);
-                    await timeout(defaultWaitForRender);
-                    for (const slice of visualBuilder.linkElements) {
-                        expect(slice.getAttribute("aria-selected")).toBe("false");
-                    }
+    describe("Focus elements tests:", () => {
+        it("focused links should have :focus-visible style", (done: DoneFn) => {
+            visualBuilder.updateRenderTimeout(dataView, () => {
+                const links: HTMLElement[] = [...visualBuilder.linkElements];
 
-                    done();
-                },
-            2,
-            );
-    });
+                links[0].focus();
+                expect(links[0].matches(':focus-visible')).toBeTrue();
 
-    it("tab between slices works", (done: DoneFn) => {
-        const tabEvent = new KeyboardEvent("keydown", { code: "Tab", bubbles: true });
-        const enterEvent = new KeyboardEvent("keydown", { code: "Enter", bubbles: true });
-        visualBuilder.updateRenderTimeout(
-            dataView,
-            async () => {
-                visualBuilder.linkElements[0].dispatchEvent(enterEvent);
-                await timeout(defaultWaitForRender);
-                expect(visualBuilder.linkElements[0].getAttribute("aria-selected")).toBe("true");
-                for (const slice of visualBuilder.linkElements) {
-                    if (slice !== visualBuilder.linkElements[0]) {
-                        expect(slice.getAttribute("aria-selected")).toBe("false");
-                    }
-                }
-
-                visualBuilder.element.dispatchEvent(tabEvent);
-                await timeout(defaultWaitForRender);
-
-                visualBuilder.linkElements[1].dispatchEvent(enterEvent);
-                await timeout(defaultWaitForRender);
-                expect(visualBuilder.linkElements[1].getAttribute("aria-selected")).toBe("true");
-                for (const slice of visualBuilder.linkElements) {
-                    if (slice !== visualBuilder.linkElements[1]) {
-                        expect(slice.getAttribute("aria-selected")).toBe("false");
-                    }
-                }
-
+                const otherLinks: HTMLElement[] = links.slice(1);
+                otherLinks.forEach((link: HTMLElement) => {
+                    expect(link.matches(':focus-visible')).toBeFalse();
+                });
                 done();
-            },
-            2,
-            );
+            });
+        });
+
+        it("focused links should have styled stroke and outline", (done: DoneFn) => {
+            visualBuilder.updateRenderTimeout(dataView, () => {
+                // defaults
+                const focusedStrokeWidth: string = "2px";
+                const focusedStrokeOpacity: string = "1";
+                const focusedOutline: string = "rgb(0, 0, 0) none 0px";
+                const strokeWidth: string = "1px";
+                const strokeOpacity: string = "0.2";
+                const outline: string = "rgb(0, 0, 0) none 0px";
+
+                const links: HTMLElement[] = [...visualBuilder.linkElements];
+
+                links[0].focus();
+
+                let linkComputedStyle: CSSStyleDeclaration = getComputedStyle(links[0]);
+                let linkStrokeWidth: string = linkComputedStyle.getPropertyValue("stroke-width");
+                let linkStrokeOpacity: string = linkComputedStyle.getPropertyValue("stroke-opacity");
+                let linkOutline: string = linkComputedStyle.getPropertyValue("outline");
+
+                expect(linkStrokeWidth).toBe(focusedStrokeWidth);
+                expect(linkStrokeOpacity).toBe(focusedStrokeOpacity);
+                expect(linkOutline).toBe(focusedOutline);
+
+                const otherLinks: HTMLElement[] = links.slice(1);
+                otherLinks.forEach((link: HTMLElement) => {
+                    linkComputedStyle = getComputedStyle(link);
+                    linkStrokeWidth = linkComputedStyle.getPropertyValue("stroke-width");
+                    linkStrokeOpacity = linkComputedStyle.getPropertyValue("stroke-opacity");
+                    linkOutline = linkComputedStyle.getPropertyValue("outline");
+
+                    expect(linkStrokeWidth).toBe(strokeWidth);
+                    expect(linkStrokeOpacity).toBe(strokeOpacity);
+                    expect(linkOutline).toBe(outline);
+                    expect(linkStrokeWidth < focusedStrokeWidth).toBeTrue();
+                });
+                done();
+            });
+        });
+
+        it("nodes should have :focus-visible style", (done: DoneFn) => {
+            visualBuilder.updateRenderTimeout(dataView, () => {
+                const nodeRects: HTMLElement[] = [...visualBuilder.nodeRectElements];
+
+                nodeRects[0].focus();
+                expect(nodeRects[0].matches(':focus-visible')).toBeTrue();
+
+                const otherNodeRects: HTMLElement[] = nodeRects.slice(1);
+                otherNodeRects.forEach((nodeRect: HTMLElement) => {
+                    expect(nodeRect.matches(':focus-visible')).toBeFalse();
+                });
+                done();
+            });
+        });
+
+        it("focused nodes should have styled stroke and outline", (done: DoneFn) => {
+            visualBuilder.updateRenderTimeout(dataView, () => {
+                // defaults
+                const focusedStrokeWidth: string = "4px";
+                const focusedOutline: string = "rgb(0, 0, 0) none 0px";
+                const strokeWidth: string = "1px";
+                const outline: string = "rgb(0, 0, 0) none 0px";
+
+                const nodeRects: HTMLElement[] = [...visualBuilder.nodeRectElements];
+
+                nodeRects[0].focus();
+
+                let nodeComputedStyle: CSSStyleDeclaration = getComputedStyle(nodeRects[0]);
+                let nodeStrokeWidth: string = nodeComputedStyle.getPropertyValue("stroke-width");
+                let nodeOutline: string = nodeComputedStyle.getPropertyValue("outline");
+                expect(nodeStrokeWidth).toBe(focusedStrokeWidth);
+                expect(nodeOutline).toBe(focusedOutline);
+
+                const otherNodeRects: HTMLElement[] = nodeRects.slice(1);
+                otherNodeRects.forEach((nodeRect: HTMLElement) => {
+                    nodeComputedStyle = getComputedStyle(nodeRect);
+                    nodeStrokeWidth = nodeComputedStyle.getPropertyValue("stroke-width");
+                    nodeOutline = nodeComputedStyle.getPropertyValue("outline");
+                    expect(nodeStrokeWidth).toBe(strokeWidth);
+                    expect(nodeOutline).toBe(outline);
+                    expect(nodeStrokeWidth < focusedStrokeWidth).toBeTrue();
+                });
+                done();
+            });
+        });
     });
 
     describe("high contrast mode test", () => {